refactor(tcp-plug-socket): split lib into more files
