+++ smarthome/crates/libs/tcp-plug-socket/src/command.rs
@@ -1,5 +1,3 @@
-use std;
-
 use std::fmt::Display;
 
 #[derive(Clone, Copy, PartialEq, PartialOrd, Eq, Ord, Debug, Default)]
