+++ smarthome/crates/libs/tcp-plug-socket/src/lib.rs
@@ -1,116 +1,19 @@
 use std::{
-    error,
-    fmt::{self},
     io::{Read, Write},
     net::{TcpStream, ToSocketAddrs},
 };
 
-pub type Result<T> = core::result::Result<T, Error>;
-pub type Error = Box<dyn std::error::Error>; // For early dev.
-
-use command::Command;
+pub use self::command::Command;
+pub use self::responses::{OddResponse, Response};
+use error::WrongResponse;
 use smarthome::devices::{Gauge, SendCommand};
 
-pub mod command;
+type Result<T> = core::result::Result<T, Error>;
+type Error = Box<dyn std::error::Error>; // For early dev.
 
-#[derive(Clone, Copy, PartialEq, PartialOrd, Eq, Ord, Debug, Default)]
-pub enum OddResponse {
-    #[default]
-    Ok,
-    Enabled,
-    Disabled,
-    Retry,
-    Power(u16),
-    MaxPower(u16),
-    Reserved([u8; 3]),
-}
-
-impl From<&[u8; 4]> for OddResponse {
-    fn from(&bytes: &[u8; 4]) -> Self {
-        match bytes {
-            [0, ..] => Self::Ok,
-            [1, ..] => Self::Enabled,
-            [2, ..] => Self::Disabled,
-            [3, ..] => Self::Retry,
-            [4, ..] => Self::Power(u16::from_be_bytes(
-                bytes[2..].try_into().unwrap_or_default(),
-            )),
-
-            [5, ..] => Self::MaxPower(u16::from_be_bytes(
-                bytes[2..].try_into().unwrap_or_default(),
-            )),
-            _ => {
-                let mut v = [0u8; 3];
-                v.clone_from_slice(&bytes[1..4]);
-                Self::Reserved(v)
-            }
-        }
-    }
-}
-
-impl From<&OddResponse> for [u8; 4] {
-    fn from(value: &OddResponse) -> Self {
-        let mut buf = [0u8; 4];
-        match value {
-            OddResponse::Ok => {}
-            OddResponse::Enabled => buf[0] = 1,
-            OddResponse::Disabled => buf[0] = 2,
-            OddResponse::Retry => buf[0] = 3,
-
-            OddResponse::Power(pw) => {
-                buf[0] = 4;
-                buf[2..].copy_from_slice(&pw.to_be_bytes());
-            }
-            OddResponse::MaxPower(mpw) => {
-                buf[0] = 5;
-                buf[2..].copy_from_slice(&mpw.to_be_bytes());
-            }
-            OddResponse::Reserved(b3) => {
-                buf[0] = 6;
-                buf[1..].copy_from_slice(b3);
-            }
-        };
-        buf
-    }
-}
-
-impl fmt::Display for OddResponse {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Ok => write!(f, "Ok"),
-            Self::Enabled => write!(f, "Enabled"),
-            Self::Disabled => write!(f, "Disabled"),
-            Self::Power(power) => write!(f, "Power: {power}"),
-            Self::MaxPower(power) => write!(f, "Max power: {power}"),
-            Self::Reserved(v) => write!(f, "Reserved {v:?}"),
-            Self::Retry => write!(f, "Retry later"),
-        }
-    }
-}
-
-#[derive(Clone, Copy, PartialEq, PartialOrd, Debug, Default)]
-pub enum Response {
-    #[default]
-    Ok,
-    Enabled,
-    Disabled,
-    Power(f32),
-    MaxPower(f32),
-    Reserved([u8; 3]),
-}
-
-impl fmt::Display for Response {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Ok => write!(f, "Ok"),
-            Self::Enabled => write!(f, "Enabled"),
-            Self::Disabled => write!(f, "Disabled"),
-            Self::Power(power) => write!(f, "Power: {power}"),
-            Self::MaxPower(power) => write!(f, "Max power: {power}"),
-            Self::Reserved(v) => write!(f, "Reserved {v:?}"),
-        }
-    }
-}
+mod command;
+mod error;
+mod responses;
 
 /// Odd TCP Smart Plug socket client- connects to smart plug via `TcpStream`
 /// let's pretend for some reason `HW` implementation can't handle floats
@@ -123,25 +26,6 @@ pub struct TcpPlugOddSocket {
 
 #[derive(Clone, Copy)]
 struct Power(f32);
-
-#[derive(Debug, Clone)]
-struct WrongResponse {
-    message: String,
-}
-
-impl WrongResponse {
-    const fn new(message: String) -> Self {
-        Self { message }
-    }
-}
-
-impl fmt::Display for WrongResponse {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "WrongResponse: {}", self.message)
-    }
-}
-
-impl error::Error for WrongResponse {}
 
 /// returns cached power usage
 impl Gauge<Power> for TcpPlugOddSocket {
